+++ tools/examples/rpg-events.js
@@ -17,7 +17,6 @@ AdRPG.setCurrentMap('event-demo');
 AdRPG.setMap(map);
 AdRPG.player.x = 6;
 AdRPG.player.y = 5;
-AdRPG.player.health = 100;
 AdRPG.player.maxHealth = 100;
 
 AdRPG.initInput();